Remove spurious zero case in shift opcode implementation
I believe when I wrote it, I wanted to guard against UB on in the
implementation of bit rotation. I forgot that destination operand
need not be the same as the source operand.

In the end, I mitigate the UB in the rol32 helper function, so the
check is just redundant.

Good thing I write those tests.